// src/sdks/Newcomputation/computationV2ApiDocs.ts
export type getComputationEmissionSourcePageProps = {
  ghg?: string | null | undefined;
  iso?: string | null | undefined;
  likeFacility?: string | null | undefined;
  likeFacilityEn?: string | null | undefined;
  likeSourceName?: string | null | undefined;
  likeSourceNameEn?: string | null | undefined;
  pageNum: number;
  pageSize: number;
  sourceCode?: string | null | undefined;
  status?: string;
  emissionStatus?: string;
  orgCode?: string;
  [key: string]: any;
};